+++ src/plugins.ts
@@ -46,6 +46,9 @@ export const pluginsLoaders = {
     interopDefault(import('eslint-plugin-css')),
   'de-morgan': () => interopDefault<EslintPlugin>(import('eslint-plugin-de-morgan')),
   depend: () => interopDefault(import('eslint-plugin-depend')),
+  'erasable-syntax-only': () =>
+    // @ts-expect-error types mismatch
+    interopDefault(import('eslint-plugin-erasable-syntax-only')),
   es: () => interopDefault(import('eslint-plugin-es-x')),
   html: () => interopDefault(import('eslint-plugin-html')),
   import: () =>